backend-django/requirements/developments.txt
astroid==2.1.0
atomicwrites==1.2.1
attrs==18.2.0
backcall==0.1.0
certifi==2018.11.29
chardet==3.0.4
colorama==0.4.1
coverage==4.5.2
decorator==4.3.0
Django==2.1.5
django-cors-middleware==1.3.1
django-extensions==2.1.4
django-filter==2.0.0
django-oauth-toolkit==1.2.0
djangorestframework==3.9.1
Faker==0.9.1
idna==2.8
ipdb==0.11
ipython==7.2.0
ipython-genutils==0.2.0
isort==4.3.4
jedi==0.13.2
lazy-object-proxy==1.3.1
Markdown==3.0.1
mccabe==0.6.1
mixer==6.1.3
more-itertools==5.0.0
oauthlib==3.0.0
parso==0.3.1
pickleshare==0.7.5
pluggy==0.8.1
prompt-toolkit==2.0.7
py==1.7.0
Pygments==2.3.1
pylint==2.2.2
pytest==4.1.1
pytest-cov==2.6.1
pytest-django==3.4.5
pytest-ipdb==0.1
python-dateutil==2.7.5
pytz==2018.9
requests==2.21.0
six==1.12.0
text-unidecode==1.2
traitlets==4.3.2
urllib3==1.24.1
wcwidth==0.1.7
wrapt==1.11.0
